% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{split_var}
\alias{split_var}
\title{Split a character variable into multiple variables}
\usage{
split_var(data, target_var, outvar_name, sep = ";", len = 200L)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{target_var}{A character variable in input data set that need to be split, like \code{product_term}.
if the name of \code{target_var} is same as \code{outvar_name}, then the name of \code{target_var} is appended with a prefix, \code{(.)}.}

\item{outvar_name}{A string which is the prefix of similar variable names, like \code{"CMDECOD"}.}

\item{sep}{A character vector that is separator delimiting collapsed values.
Default: \verb{;}.}

\item{len}{A scalar integer(a length-one numeric) to assign the length that every variable shouldn't exceed.}
}
\value{
A tibble including a series of variable calling \code{outvar_name}, which is invisible object.
}
\description{
\code{split_var} separates a character variable with a length exceeding \code{length} into multiple variables
with the longest not exceeding \code{length} with corresponding delimiter, which is suitable for data manipulation.
}
\examples{

test_df <- data.frame(
  id = 1:2,
  text = c(
    paste(strrep("a", 150), ";", strrep("a", 50), strrep("a", 130), sep = ";"),
    paste(strrep("a", 150), strrep("a", 30), strrep("a", 150), strrep("a", 80), sep = ";")
  )
)


new_test_df <- test_df \%>\% dso_split_var(text, "text", "\\\\;")

}
